+++ client/src/MobicoopBundle/Resources/translations/components/base/MMessageBtn/index.js
@@ -1,0 +1,2 @@
+export {default as messages_fr} from './MMessageBtn_fr.json';
+export {default as messages_en} from './MMessageBtn_en.json';