+++ client/src/MobicoopBundle/Resources/translations/components/user/mailbox/ThreadDirect/index.js
@@ -1,0 +1,2 @@
+export {default as messages_fr} from './ThreadDirect_fr.json';
+export {default as messages_en} from './ThreadDirect_en.json';