+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Release 1.22.0
+* New contact recipients system
+* Number of unread messages
+
 # Release 1.20.0